Rename row generation module to use snake case
