+++ row_generation/complib_composition_reader.py
@@ -2,8 +2,8 @@ from typing import List
 
 import requests
 
-from RowGeneration.Helpers import convert_bell_string
-from RowGeneration.RowGenerator import RowGenerator
+from .helpers import convert_bell_string
+from .row_generator import RowGenerator
 from bell import Bell
