+++ row_generation/go_and_stop_calling_generator.py
@@ -1,8 +1,8 @@
 from random import random
 from typing import List
 
-from Calls import Calls
-from RowGeneration.RowGenerator import RowGenerator
+import calls
+from .row_generator import RowGenerator
 from bell import Bell
 from tower import RingingRoomTower
 
@@ -17,12 +17,12 @@ class GoAndStopCallingGenerator(RowGenerator):
 
     def next_row(self, is_handstroke: bool):
         if not self.called_go and not is_handstroke and random.choices([True, False], [1, 3]):
-            self.tower.make_call(Calls.Go)
+            self.tower.make_call(calls.GO)
         return super(GoAndStopCallingGenerator, self).next_row(is_handstroke)
 
     def _gen_row(self, previous_row: List[Bell], is_handstroke: bool, index: int) -> List[Bell]:
         next_row = self.generator._gen_row(previous_row, is_handstroke, index)
         if next_row == self.rounds():
-            self.tower.make_call(Calls.ThatsAll)
+            self.tower.make_call(calls.THATS_ALL)
 
         return next_row
