+++ row_generation/helpers.py
